App/requirements.txt
aniso8601==9.0.1
APScheduler==3.9.1.post1
certifi==2021.10.8
charset-normalizer
click
colorama==0.4.5
Flask
Flask-APScheduler==1.12.4
Flask-RESTful==0.3.9
idna==3.4
importlib-metadata==4.8.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.19.5
packaging==21.3
paho-mqtt==1.6.1
pandas==1.1.5
psutil==5.9.4
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
requests==2.27.1
resources==0.0.1
six==1.16.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
websocket-client==1.3.1
Werkzeug==2.0.3
